Simplify Composio integration structure
