+++ src/components/content/DirectoryHub.vue
@@ -711,49 +711,28 @@
 import { computed, onMounted, ref, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import {
-  getDirectoryComposioStatus,
   getMethodCatalog,
   installDirectoryPlugin,
-  installDirectoryComposioCli,
-  listDirectoryComposioConnectors,
   listDirectoryApps,
   listDirectoryMcpServers,
   listDirectoryPlugins,
-  logoutDirectoryComposioCli,
-  readDirectoryComposioConnector,
   readDirectoryPlugin,
   reloadDirectoryMcpServers,
   setDirectoryAppEnabled,
   setDirectoryPluginEnabled,
-  startDirectoryComposioCliLogin,
-  startDirectoryComposioLogin,
   startDirectoryMcpLogin,
   uninstallDirectoryPlugin,
-  uploadFile,
   type DirectoryAppInfo,
-  type DirectoryComposioConnection,
-  type DirectoryComposioConnector,
-  type DirectoryComposioConnectorDetail,
-  type DirectoryComposioStatus,
   type DirectoryMcpServerStatus,
   type DirectoryPluginAppSummary,
   type DirectoryPluginDetail,
   type DirectoryPluginSummary,
 } from '../../api/codexGateway'
-import { HARDCODED_COMPOSIO_CONNECTORS } from './composioConnectorCatalog'
-import {
-  buildComposioConnectorDocument,
-  composioConnectorDocumentFileName,
-  mergeComposioConnectors,
-} from './composioComposerSuggestions'
-import { sortComposioConnectors, type DirectorySortMode } from './directoryHubUtils'
+import { type DirectorySortMode } from './directoryHubUtils'
 import SkillsHub from './SkillsHub.vue'
+import { DEFAULT_COMPOSIO_DASHBOARD_URL, useComposioDirectory } from './useComposioDirectory'
 
 type DirectoryTab = 'plugins' | 'apps' | 'composio' | 'skills'
-const COMPOSIO_PAGE_LIMIT = 50
-const DEFAULT_COMPOSIO_DASHBOARD_URL = 'https://dashboard.composio.dev/'
-const COMPOSIO_AUTH_POLL_INTERVAL_MS = 2_000
-const COMPOSIO_AUTH_POLL_TIMEOUT_MS = 120_000
 
 const POPULAR_LIMIT = 100
 const POPULAR_APP_NAME_BONUSES: Array<[RegExp, number]> = [
@@ -849,46 +828,27 @@ const methodSet = ref<Set<string>>(new Set())
 const methodsLoaded = ref(false)
 const plugins = ref<DirectoryPluginSummary[]>([])
 const apps = ref<DirectoryAppInfo[]>([])
-const composioStatus = ref<DirectoryComposioStatus | null>(null)
-const composioConnectors = ref<DirectoryComposioConnector[]>([])
-const composioNextCursor = ref<string | null>(null)
-const composioTotal = ref(0)
-const composioVisibleLimit = ref(COMPOSIO_PAGE_LIMIT)
 const mcpServers = ref<DirectoryMcpServerStatus[]>([])
 const pluginSortMode = ref<DirectorySortMode>('popular')
 const appSortMode = ref<DirectorySortMode>('popular')
-const composioSortMode = ref<DirectorySortMode>('popular')
 const pluginSearchQuery = ref('')
 const appSearchQuery = ref('')
-const composioSearchQuery = ref('')
 const isLoadingPlugins = ref(false)
 const isLoadingApps = ref(false)
-const isLoadingComposio = ref(false)
 const isLoadingMcps = ref(false)
 const isReloadingMcps = ref(false)
 const isManualRefreshInFlight = ref(false)
 const isMcpSectionOpen = ref(true)
 const pluginError = ref('')
 const appError = ref('')
-const composioError = ref('')
 const mcpError = ref('')
-const lastRouteComposioConnectorSlug = ref('')
 const selectedPlugin = ref<DirectoryPluginSummary | null>(null)
 const selectedPluginDetail = ref<DirectoryPluginDetail | null>(null)
 const isPluginDetailOpen = ref(false)
 const isLoadingPluginDetail = ref(false)
 const pluginDetailError = ref('')
-const selectedComposioDetail = ref<DirectoryComposioConnectorDetail | null>(null)
-const isComposioDetailOpen = ref(false)
-const isLoadingComposioDetail = ref(false)
-const composioDetailError = ref('')
-const isInstallingComposio = ref(false)
-const isStartingComposioLogin = ref(false)
-const isLoggingOutComposio = ref(false)
 const isPluginActionInFlight = ref(false)
 const appActionId = ref('')
-const composioActionSlug = ref('')
-const composioTryUploadSlug = ref('')
 const installAuthApps = ref<DirectoryPluginAppSummary[]>([])
 const mcpLoginServerName = ref('')
 const expandedMcpNames = ref<Set<string>>(new Set())
@@ -905,6 +865,54 @@ const supportsMcps = computed(() => !methodsLoaded.value || methodSet.value.has(
 const supportsMcpReload = computed(() => methodSet.value.has('config/mcpServer/reload'))
 const supportsMcpLogin = computed(() => methodSet.value.has('mcpServer/oauth/login'))
 const isTryActionInFlight = computed(() => (props.tryInFlightKey ?? '').length > 0)
+const {
+  composioStatus,
+  composioConnectors,
+  composioTotal,
+  composioSortMode,
+  composioSearchQuery,
+  isLoadingComposio,
+  composioError,
+  selectedComposioDetail,
+  isComposioDetailOpen,
+  isLoadingComposioDetail,
+  composioDetailError,
+  isInstallingComposio,
+  isStartingComposioLogin,
+  isLoggingOutComposio,
+  composioActionSlug,
+  composioTryUploadSlug,
+  visibleComposioConnectors,
+  hasMoreComposioConnectors,
+  composioWorkspaceSummary,
+  composioLogoSrc,
+  composioMetaLabel,
+  composioPrimaryActionLabel,
+  composioConnectionStatusLabel,
+  composioConnectionStatusClass,
+  resetComposioVisibleLimit,
+  composioTryKey,
+  canTryComposio,
+  tryComposio,
+  loadComposio,
+  loadMoreComposio,
+  openComposioDetail,
+  closeComposioDetail,
+  openRouteComposioConnector,
+  resetRouteComposioConnectorSlug,
+  runComposioPrimaryAction,
+  startComposioCliLogin,
+  installComposioCli,
+  logoutComposioCli,
+} = useComposioDirectory({
+  activeTab,
+  isTryActionInFlight,
+  getRouteConnectorSlug: composioConnectorSlugFromRoute,
+  localAssetSrc,
+  openExternalUrl,
+  showToast,
+  emitTryItem: (payload) => emit('try-item', payload),
+})
 const selectedPluginDescription = computed(() =>
   selectedPluginDetail.value?.description ||
   selectedPluginDetail.value?.summary.longDescription ||
@@ -938,29 +946,8 @@ const selectedPluginInstallUnavailable = computed(() =>
 )
 const visiblePlugins = computed(() => limitPopularRows(sortPlugins(filterPlugins(plugins.value, pluginSearchQuery.value), pluginSortMode.value), pluginSortMode.value, pluginSearchQuery.value))
 const visibleApps = computed(() => limitPopularApps(sortApps(filterApps(apps.value, appSearchQuery.value), appSortMode.value), appSortMode.value, appSearchQuery.value))
-const filteredComposioConnectors = computed(() =>
-  sortComposioConnectors(
-    filterComposioConnectors(composioConnectors.value, composioSearchQuery.value),
-    composioSortMode.value,
-    composioSearchQuery.value,
-  ),
-)
-const visibleComposioConnectors = computed(() => filteredComposioConnectors.value.slice(0, composioVisibleLimit.value))
 const visibleMcpServers = computed(() => sortMcpServers(mcpServers.value, 'popular'))
-const hasMoreComposioConnectors = computed(() =>
-  visibleComposioConnectors.value.length < filteredComposioConnectors.value.length || composioNextCursor.value !== null,
-)
 const mcpStatusByName = computed(() => new Map(mcpServers.value.map((server) => [server.name, server])))
-const composioWorkspaceSummary = computed(() => {
-  const status = composioStatus.value
-  if (!status) return 'Composio CLI shares the login and connections from this machine.'
-  const parts = [
-    status.email || status.defaultOrgName,
-    status.defaultOrgId ? `org ${status.defaultOrgId}` : '',
-    status.baseUrl || '',
-  ].filter(Boolean)
-  return parts.join(' · ') || 'Composio CLI shares the login and connections from this machine.'
-})
 
 function normalizeSearch(value: string): string {
   return value.trim().toLowerCase()
@@ -1109,16 +1096,6 @@ function filterApps(rows: DirectoryAppInfo[], query: string): DirectoryAppInfo[]
   ], query))
 }
 
-function filterComposioConnectors(rows: DirectoryComposioConnector[], query: string): DirectoryComposioConnector[] {
-  return rows.filter((connector) => includesSearch([
-    connector.name,
-    connector.slug,
-    connector.description,
-    ...connector.authModes,
-    ...connector.connectionStatuses,
-  ], query))
-}
-
 function pluginPopularScore(plugin: DirectoryPluginSummary): number {
   return (
     (plugin.installed ? 500 : 0) +
@@ -1195,63 +1172,6 @@ function appLogoSrc(app: DirectoryAppInfo): string {
   return localAssetSrc(app.logoUrlDark || app.logoUrl)
 }
 
-function composioLogoSrc(connector: DirectoryComposioConnector): string {
-  return localAssetSrc(connector.logoUrl)
-}
-
-function composioMetaLabel(connector: DirectoryComposioConnector): string {
-  if (connector.activeCount > 0) {
-    return `${connector.activeCount} connected ${connector.activeCount === 1 ? 'account' : 'accounts'}`
-  }
-  if (connector.isNoAuth) return 'No auth required'
-  if (connector.connectionStatuses.length > 0) return connector.connectionStatuses.join(', ')
-  return connector.authModes.join(', ') || 'Connection required'
-}
-
-function composioHasUsableConnection(connector: DirectoryComposioConnector): boolean {
-  return connector.isNoAuth || connector.activeCount > 0
-}
-
-function composioPrimaryActionLabel(connector: DirectoryComposioConnector): string {
-  if (connector.isNoAuth) return ''
-  if (connector.activeCount > 0) return 'Manage'
-  if (connector.totalConnections > 0) return 'Reconnect'
-  return 'Connect'
-}
-
-function composioConnectionStatusLabel(status: string): string {
-  const normalized = status.trim().toUpperCase()
-  if (normalized === 'ACTIVE') return 'Active'
-  if (normalized === 'EXPIRED') return 'Expired'
-  if (normalized === 'FAILED') return 'Failed'
-  if (normalized === 'INITIATED') return 'Pending'
-  return normalized || 'Unknown'
-}
-
-function composioConnectionStatusClass(status: string): string {
-  const normalized = status.trim().toUpperCase()
-  if (normalized === 'ACTIVE') return 'is-ok'
-  if (normalized === 'INITIATED') return 'is-warning'
-  if (normalized === 'EXPIRED' || normalized === 'FAILED') return 'is-error'
-  return 'is-muted'
-}
-
-function buildLocalComposioDetail(
-  connector: DirectoryComposioConnector,
-  connections: DirectoryComposioConnection[] = [],
-): DirectoryComposioConnectorDetail {
-  return {
-    connector,
-    connections,
-    tools: [],
-    dashboardUrl: composioStatus.value?.webUrl || DEFAULT_COMPOSIO_DASHBOARD_URL,
-  }
-}
-
-function resetComposioVisibleLimit(): void {
-  composioVisibleLimit.value = COMPOSIO_PAGE_LIMIT
-}
-
 function appTryKey(app: DirectoryAppInfo): string {
   return `app:${app.id}:`
 }
@@ -1269,10 +1189,6 @@ function pluginTryKey(plugin: DirectoryPluginSummary): string {
   return `plugin:${plugin.name}:`
 }
 
-function composioTryKey(slug: string): string {
-  return `composio:${slug}:`
-}
-
 function tryPlugin(plugin: DirectoryPluginSummary): void {
   if (isTryActionInFlight.value) return
   emit('try-item', {
@@ -1282,77 +1198,10 @@ function tryPlugin(plugin: DirectoryPluginSummary): void {
   })
 }
 
-function canTryComposio(connector: DirectoryComposioConnector): boolean {
-  return composioHasUsableConnection(connector)
-}
-
-function buildComposioTryPrompt(connector: DirectoryComposioConnector, connections: DirectoryComposioConnection[] = []): string {
-  const firstActive = connections.find((connection) => connection.status === 'ACTIVE' && !connection.isDisabled)
-  const accountHint = firstActive?.wordId
-    ? ` If there are multiple accounts, prefer \`${firstActive.wordId}\`.`
-    : ''
-  return `Use the attached ${connector.name} Composio connector documentation for this request. Start by listing what it can do here, mention the current connection status, and suggest one safe action I can run now.${accountHint}`
-}
-
-async function buildComposioTryAttachment(
-  connector: DirectoryComposioConnector,
-  connections: DirectoryComposioConnection[] = [],
-): Promise<{ label: string; path: string; fsPath: string } | null> {
-  let detail: DirectoryComposioConnectorDetail | null = null
-  try {
-    detail = await readDirectoryComposioConnector(connector.slug)
-  } catch {
-    detail = buildLocalComposioDetail(connector, connections)
-  }
-
-  const fileName = composioConnectorDocumentFileName(connector)
-  const document = buildComposioConnectorDocument(connector, detail)
-  const file = new File([document], fileName, {
-    type: 'text/markdown',
-    lastModified: Date.now(),
-  })
-  const serverPath = await uploadFile(file)
-  if (!serverPath) return null
-  return { label: fileName, path: serverPath, fsPath: serverPath }
-}
-
-async function tryComposio(connector: DirectoryComposioConnector, connections: DirectoryComposioConnection[] = []): Promise<void> {
-  if (isTryActionInFlight.value || composioTryUploadSlug.value) return
-  composioTryUploadSlug.value = connector.slug
-  let fileAttachment: { label: string; path: string; fsPath: string } | null = null
-  try {
-    fileAttachment = await buildComposioTryAttachment(connector, connections)
-  } finally {
-    composioTryUploadSlug.value = ''
-  }
-  if (!fileAttachment) {
-    showToast(`Failed to attach ${connector.name} documentation.`, 'error')
-    return
-  }
-  emit('try-item', {
-    kind: 'composio',
-    name: connector.slug,
-    displayName: connector.name,
-    prompt: buildComposioTryPrompt(connector, connections),
-    fileAttachments: [fileAttachment],
-  })
-}
-
 function openExternalUrl(rawUrl: string): void {
   const url = rawUrl.trim()
   if (!/^https?:\/\//i.test(url)) return
   window.location.assign(url)
-}
-
-function openExternalUrlInNewTab(rawUrl: string): boolean {
-  const url = rawUrl.trim()
-  if (!/^https?:\/\//i.test(url)) return false
-  const tab = window.open(url, '_blank', 'noopener,noreferrer')
-  return Boolean(tab)
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms))
 }
 
 function openFirstAppLoginIfNeeded(apps: DirectoryPluginAppSummary[]): boolean {
@@ -1404,63 +1253,6 @@ async function loadApps(): Promise<void> {
     appError.value = error instanceof Error ? error.message : 'Failed to load apps'
   } finally {
     isLoadingApps.value = false
-  }
-}
-
-async function loadComposio(): Promise<void> {
-  isLoadingComposio.value = true
-  composioError.value = ''
-  try {
-    const statusPromise = getDirectoryComposioStatus()
-    const pagePromise = listDirectoryComposioConnectors('', null, COMPOSIO_PAGE_LIMIT)
-      .then((page) => ({ page, error: null as Error | null }))
-      .catch((error: unknown) => ({
-        page: null,
-        error: error instanceof Error ? error : new Error('Failed to load Composio connectors'),
-      }))
-    const status = await statusPromise
-    composioStatus.value = status
-    resetComposioVisibleLimit()
-    if (!status.available || !status.authenticated) {
-      composioConnectors.value = HARDCODED_COMPOSIO_CONNECTORS
-      composioNextCursor.value = null
-      composioTotal.value = HARDCODED_COMPOSIO_CONNECTORS.length
-    } else {
-      const { page, error } = await pagePromise
-      if (error || !page) throw error ?? new Error('Failed to load Composio connectors')
-      composioConnectors.value = mergeComposioConnectors(HARDCODED_COMPOSIO_CONNECTORS, page.data)
-      composioNextCursor.value = page.nextCursor
-      composioTotal.value = Math.max(page.total, composioConnectors.value.length)
-    }
-  } catch (error) {
-    composioError.value = error instanceof Error ? error.message : 'Failed to load Composio connectors'
-    composioConnectors.value = HARDCODED_COMPOSIO_CONNECTORS
-    composioNextCursor.value = null
-    composioTotal.value = HARDCODED_COMPOSIO_CONNECTORS.length
-  } finally {
-    isLoadingComposio.value = false
-  }
-  await openRouteComposioConnector()
-}
-
-async function loadMoreComposio(): Promise<void> {
-  if (!hasMoreComposioConnectors.value) return
-  if (!composioNextCursor.value) {
-    composioVisibleLimit.value += COMPOSIO_PAGE_LIMIT
-    return
-  }
-  if (!composioStatus.value?.available || !composioStatus.value.authenticated) return
-  isLoadingComposio.value = true
-  try {
-    const page = await listDirectoryComposioConnectors('', composioNextCursor.value, COMPOSIO_PAGE_LIMIT)
-    composioConnectors.value = mergeComposioConnectors(composioConnectors.value, page.data)
-    composioNextCursor.value = page.nextCursor
-    composioTotal.value = Math.max(page.total, composioConnectors.value.length)
-    composioVisibleLimit.value += COMPOSIO_PAGE_LIMIT
-  } catch (error) {
-    composioError.value = error instanceof Error ? error.message : 'Failed to load more Composio connectors'
-  } finally {
-    isLoadingComposio.value = false
   }
 }
 
@@ -1557,170 +1349,6 @@ async function openFirstMcpLoginIfNeeded(detail: DirectoryPluginDetail): Promise
 
 function closePluginDetail(): void {
   isPluginDetailOpen.value = false
-}
-
-async function openComposioDetail(slug: string): Promise<void> {
-  isComposioDetailOpen.value = true
-  isLoadingComposioDetail.value = true
-  composioDetailError.value = ''
-  selectedComposioDetail.value = null
-  try {
-    const local = composioConnectors.value.find((connector) => connector.slug === slug)
-      ?? HARDCODED_COMPOSIO_CONNECTORS.find((connector) => connector.slug === slug)
-    if (!composioStatus.value?.available || !composioStatus.value?.authenticated) {
-      if (!local) throw new Error(`Unknown Composio connector: ${slug}`)
-      selectedComposioDetail.value = buildLocalComposioDetail(local)
-      return
-    }
-    selectedComposioDetail.value = await readDirectoryComposioConnector(slug)
-  } catch (error) {
-    composioDetailError.value = error instanceof Error ? error.message : 'Failed to load Composio connector'
-  } finally {
-    isLoadingComposioDetail.value = false
-  }
-}
-
-function closeComposioDetail(): void {
-  isComposioDetailOpen.value = false
-}
-
-function updateComposioConnectorRow(connector: DirectoryComposioConnector): void {
-  composioConnectors.value = mergeComposioConnectors(composioConnectors.value, [connector])
-  composioTotal.value = Math.max(composioTotal.value, composioConnectors.value.length)
-}
-
-async function openRouteComposioConnector(): Promise<void> {
-  if (activeTab.value !== 'composio') return
-  if (isLoadingComposio.value) return
-  const slug = composioConnectorSlugFromRoute()
-  if (!slug || lastRouteComposioConnectorSlug.value === slug) return
-  composioSearchQuery.value = slug
-  await openComposioDetail(slug)
-  lastRouteComposioConnectorSlug.value = slug
-}
-
-async function startComposioConnect(connector: DirectoryComposioConnector): Promise<void> {
-  composioActionSlug.value = connector.slug
-  try {
-    const result = await startDirectoryComposioLogin(connector.slug)
-    if (!result.redirectUrl) {
-      showToast(`No login URL returned for ${connector.name}`, 'error')
-      return
-    }
-    if (!openExternalUrlInNewTab(result.redirectUrl)) {
-      showToast(`Popup blocked. Allow popups to connect ${connector.name}.`, 'error')
-      return
-    }
-    showToast(`Waiting for ${connector.name} connection...`)
-    await waitForComposioConnectorConnection(connector.slug)
-    showToast(`${connector.name} connected`)
-  } catch (error) {
-    showToast(error instanceof Error ? error.message : `Failed to connect ${connector.name}`, 'error')
-  } finally {
-    composioActionSlug.value = ''
-  }
-}
-
-async function runComposioPrimaryAction(connector: DirectoryComposioConnector): Promise<void> {
-  if (connector.activeCount > 0 && composioStatus.value?.webUrl) {
-    openExternalUrl(composioStatus.value.webUrl)
-    return
-  }
-  await startComposioConnect(connector)
-}
-
-async function startComposioCliLogin(): Promise<void> {
-  isStartingComposioLogin.value = true
-  const loginTab = window.open('about:blank', '_blank')
-  if (loginTab) {
-    loginTab.opener = null
-  }
-  try {
-    const result = await startDirectoryComposioCliLogin()
-    if (result.loginUrl && loginTab) {
-      loginTab.location.href = result.loginUrl
-    } else if (result.loginUrl) {
-      showToast('Popup blocked. Allow popups to login to Composio.', 'error')
-      return
-    } else {
-      loginTab?.close()
-    }
-    showToast('Waiting for Composio login...')
-    await waitForComposioLogin()
-    showToast('Composio CLI logged in')
-  } catch (error) {
-    loginTab?.close()
-    showToast(error instanceof Error ? error.message : 'Failed to start Composio login', 'error')
-  } finally {
-    isStartingComposioLogin.value = false
-  }
-}
-
-async function installComposioCli(): Promise<void> {
-  isInstallingComposio.value = true
-  try {
-    await installDirectoryComposioCli()
-    showToast('Composio CLI installed')
-    await loadComposio()
-  } catch (error) {
-    showToast(error instanceof Error ? error.message : 'Failed to install Composio CLI', 'error')
-  } finally {
-    isInstallingComposio.value = false
-  }
-}
-
-async function logoutComposioCli(): Promise<void> {
-  isLoggingOutComposio.value = true
-  try {
-    await logoutDirectoryComposioCli()
-    closeComposioDetail()
-    showToast('Composio CLI logged out')
-    await loadComposio()
-  } catch (error) {
-    showToast(error instanceof Error ? error.message : 'Failed to logout Composio CLI', 'error')
-  } finally {
-    isLoggingOutComposio.value = false
-  }
-}
-
-async function waitForComposioLogin(): Promise<void> {
-  const deadline = Date.now() + COMPOSIO_AUTH_POLL_TIMEOUT_MS
-  while (Date.now() < deadline) {
-    const status = await getDirectoryComposioStatus(true)
-    composioStatus.value = status
-    if (status.available && status.authenticated) {
-      await loadComposio()
-      return
-    }
-    await sleep(COMPOSIO_AUTH_POLL_INTERVAL_MS)
-  }
-  await loadComposio()
-  throw new Error('Timed out waiting for Composio login to complete')
-}
-
-async function waitForComposioConnectorConnection(slug: string): Promise<void> {
-  const deadline = Date.now() + COMPOSIO_AUTH_POLL_TIMEOUT_MS
-  while (Date.now() < deadline) {
-    let connector = composioConnectors.value.find((row) => row.slug === slug)
-    try {
-      const detail = await readDirectoryComposioConnector(slug, true)
-      connector = detail.connector
-      updateComposioConnectorRow(detail.connector)
-      if (isComposioDetailOpen.value && selectedComposioDetail.value?.connector.slug === slug) {
-        selectedComposioDetail.value = detail
-      }
-    } catch {
-      // Keep polling; the connector can fail until the external auth finishes.
-    }
-    if (connector && (connector.activeCount > 0 || connector.isNoAuth)) {
-      await loadComposio()
-      updateComposioConnectorRow(connector)
-      return
-    }
-    await sleep(COMPOSIO_AUTH_POLL_INTERVAL_MS)
-  }
-  await loadComposio()
-  throw new Error('Timed out waiting for Composio connector connection to complete')
 }
 
 async function installSelectedPlugin(): Promise<void> {
@@ -1833,7 +1461,7 @@ watch(() => route.query.tab, () => {
 })
 watch(() => route.query.connector, () => {
   if (route.name !== 'skills' || activeTab.value !== 'composio') return
-  lastRouteComposioConnectorSlug.value = ''
+  resetRouteComposioConnectorSlug()
   void openRouteComposioConnector()
 })
 watch(composioSearchQuery, () => {
